+++ packages/column-views/src/correlation-chart/stories/correlation-chart.stories.ts
@@ -4,14 +4,15 @@ import { useCorrelationLine } from "./utils";
 import {
   ColumnCorrelationMap,
   ColumnCorrelationProvider,
-  fetchUnits,
-  MacrostratDataProvider,
   MergeSectionsMode,
   useCorrelationMapStore,
-  useMacrostratBaseURL,
-  useMacrostratFetch,
 } from "../..";
 import { hyperStyled } from "@macrostrat/hyper";
+import {
+  MacrostratDataProvider,
+  fetchUnits,
+  useMacrostratFetch,
+} from "@macrostrat/data-provider";
 
 import styles from "./stories.module.sass";
 import { CorrelationChart, CorrelationChartProps } from "../main";
@@ -19,6 +20,7 @@ import { ErrorBoundary, useAsyncMemo } from "@macrostrat/ui-components";
 import { OverlaysProvider } from "@blueprintjs/core";
 import { EnvironmentColoredUnitComponent } from "../../units";
 import { scaleLinear, scalePow } from "d3-scale";
+import { MacrostratInteractionProvider } from "@macrostrat/data-components";
 
 const mapboxToken = import.meta.env.VITE_MAPBOX_API_TOKEN;
 
@@ -35,29 +37,34 @@ function CorrelationStoryUI({
   projectID,
   ...rest
 }: any) {
+  const domain = "https://dev.macrostrat.org";
   return h(
     MacrostratDataProvider,
-    { baseURL: "https://dev.macrostrat.org/api/v2" },
+    { baseURL: domain + "/api/v2" },
     h(
-      ColumnCorrelationProvider,
-      {
-        focusedLine,
-        columns: null,
-        projectID,
-        onSelectColumns(cols, line) {
-          setFocusedLine(line);
+      MacrostratInteractionProvider,
+      { linkDomain: domain },
+      h(
+        ColumnCorrelationProvider,
+        {
+          focusedLine,
+          columns: null,
+          projectID,
+          onSelectColumns(cols, line) {
+            setFocusedLine(line);
+          },
         },
-      },
-      h("div.correlation-ui", [
-        h("div.correlation-container", h(CorrelationDiagramWrapper, rest)),
-        h("div.right-column", [
-          h(ColumnCorrelationMap, {
-            accessToken: mapboxToken,
-            className: "correlation-map",
-            //showLogo: false,
-          }),
+        h("div.correlation-ui", [
+          h("div.correlation-container", h(CorrelationDiagramWrapper, rest)),
+          h("div.right-column", [
+            h(ColumnCorrelationMap, {
+              accessToken: mapboxToken,
+              className: "correlation-map",
+              //showLogo: false,
+            }),
+          ]),
         ]),
-      ]),
+      ),
     ),
   );
 }