+++ packages/column-views/src/data-provider/index.ts
@@ -1,5 +1,3 @@
-export * from "./fetch";
-export * from "./base";
 export * from "./column-nav";
 export * from "./store";
 export * from "./unit-selection";