+++ packages/column-views/src/data-provider/unit-selection.ts
@@ -1,15 +1,15 @@
-import { BaseUnit, UnitLong } from "@macrostrat/api-types";
+import type { BaseUnit, UnitLong } from "@macrostrat/api-types";
 import { useKeyHandler } from "@macrostrat/ui-components";
 import { useEffect, useRef, useCallback, useMemo } from "react";
 import type { RectBounds, IUnit } from "../units/types";
 import { atom } from "jotai";
-import { columnUnitsMapAtom, scope } from "./core";
-import { ColumnData } from "@macrostrat/column-views";
+import { scope, columnUnitsMapAtom } from "./store";
 import {
   AgeRangeQuantifiedDifference,
   ageRangeQuantifiedDifference,
   AgeRangeRelationship,
 } from "@macrostrat/stratigraphy-utils";
+import type { ColumnData } from "@macrostrat/data-provider";
 
 type UnitSelectDispatch = (
   unit: number | BaseUnit | null,
@@ -69,7 +69,7 @@ const selectedUnitAtom = atom(
     target: HTMLElement | null = null,
   ): BaseUnit | null => {
     if (!get(allowUnitSelectionAtom)) {
-      throw new Error("Unit selection is disabled.");
+      console.error("Unit selection is disabled.");
     }
 
     let unitID: number | null;