+++ packages/column-views/src/unit-details/panel.ts
@@ -2,7 +2,7 @@ import hyper from "@macrostrat/hyper";
 import styles from "./panel.module.sass";
 import { JSONView } from "@macrostrat/ui-components";
 import { Button, ButtonGroup } from "@blueprintjs/core";
-import { ReactNode, useMemo, useState } from "react";
+import { ReactNode, useCallback, useMemo, useState } from "react";
 import {
   DataField,
   EnvironmentsList,
@@ -13,13 +13,21 @@ import {
   LithologyTagFeature,
   Parenthetical,
   Value,
+  MacrostratInteractionManager,
+  useInteractionProps,
+  ItemInteractionProps,
+  useInteractionManager,
+  MacrostratItemIdentifier,
+  MacrostratInteractionProvider,
+  isClickable,
 } from "@macrostrat/data-components";
+import { useColumnUnitsMap } from "../data-provider";
 import {
-  useColumnUnitsMap,
   useMacrostratColumnInfo,
   useMacrostratData,
   useMacrostratDefs,
-} from "../data-provider";
+  useStratNames,
+} from "@macrostrat/data-provider";
 import type {
   Environment,
   UnitLong,
@@ -42,6 +50,7 @@ export interface UnitDetailsPanelProps {
   lithologyFeatures?: Set<LithologyTagFeature>;
   onSelectUnit?: (unitID: number) => void;
   onClickItem?: MacrostratItemClickHandler;
+  interactionManager?: MacrostratInteractionManager;
 }
 
 export function UnitDetailsPanel({
@@ -71,7 +80,6 @@ export function UnitDetailsPanel({
       features,
       lithologyFeatures,
       onClickItem,
-      onSelectUnit,
     });
   }
 
@@ -94,7 +102,7 @@ export function UnitDetailsPanel({
     ]);
   }
 
-  return h("div.unit-details-panel", { className }, [
+  const main = h("div.unit-details-panel", { className }, [
     h(LegendPanelHeader, {
       onClose,
       title,
@@ -104,6 +112,36 @@ export function UnitDetailsPanel({
     }),
     h("div.unit-details-content-holder", content),
   ]);
+
+  /** Handle unit selection clicks */
+  const clickHandlerForItem = useMemo(() => {
+    if (onSelectUnit == null && onClickItem == null) return null;
+    return (item: MacrostratItemIdentifier) => {
+      if ("unit_id" in item && !("col_id" in item)) {
+        // We are selecting a unit within the column
+        return (event: MouseEvent) => {
+          onSelectUnit(item.unit_id);
+          // Don't allow event to propagate further (e.g., to open a link)
+          event.preventDefault();
+        };
+      }
+      return undefined;
+    };
+  }, [onSelectUnit, onClickItem]);
+
+  if (clickHandlerForItem != null) {
+    // We wrap this in a MacrostratInteractionManager to handle unit selection
+    return h(
+      MacrostratInteractionProvider,
+      {
+        clickHandlerForItem,
+        inherit: true,
+      },
+      main,
+    );
+  }
+
+  return main;
 }
 
 export function LegendPanelHeader({
@@ -162,7 +200,6 @@ export type MacrostratItemClickHandler = (
 
 function UnitDetailsContent({
   unit,
-  onSelectUnit,
   lithologyFeatures = new Set([
     LithologyTagFeature.Proportion,
     LithologyTagFeature.Attributes,
@@ -175,7 +212,6 @@ function UnitDetailsContent({
   getItemHref,
 }: {
   unit: UnitLong;
-  onSelectUnit?: (unitID: number) => void;
   lithologyFeatures?: Set<LithologyTagFeature>;
   features?: Set<UnitDetailsFeature>;
   onClickItem?: MacrostratItemClickHandler;
@@ -243,12 +279,10 @@ function UnitDetailsContent({
       col_id: unit.col_id,
     }),
     thicknessOrHeightRange,
-    h.if(lithologies)(LithologyList, {
+    h.if(lithologies != null)(LithologyList, {
       label: "Lithology",
       lithologies,
       features: lithologyFeatures,
-      getItemHref,
-      onClickItem,
     }),
     h(AgeField, { unit }, [
       h(Parenthetical, h(Duration, { value: unit.b_age - unit.t_age })),
@@ -257,10 +291,8 @@ function UnitDetailsContent({
         onClickItem,
       }),
     ]),
-    h.if(environments)(EnvironmentsList, {
+    h.if(environments != null)(EnvironmentsList, {
       environments,
-      onClickItem,
-      getItemHref,
     }),
     h.if(unit.strat_name_id != null)(StratNameField, {
       strat_name_id: unit.strat_name_id,
@@ -272,7 +304,6 @@ function UnitDetailsContent({
         { label: "Above" },
         h(UnitIDList, {
           units: unit.units_above,
-          onSelectUnit,
           showNames: true,
         }),
       ),
@@ -281,7 +312,6 @@ function UnitDetailsContent({
         { label: "Below" },
         h(UnitIDList, {
           units: unit.units_below,
-          onSelectUnit,
           showNames: true,
         }),
       ),
@@ -333,27 +363,35 @@ export function ReferencesField({ refs, className = null, ...rest }) {
   );
 }
 
-function StratNameField({
-  strat_name_id,
-  onClickItem,
-}: {
-  strat_name_id: number;
-  onClickItem?: (event: MouseEvent, item: { strat_name_id: number }) => void;
-}) {
-  const stratNames = useMemo(() => [strat_name_id], [strat_name_id]);
-  const data = useMacrostratData("strat_names", stratNames);
-  const stratNameData = data?.[0];
+function useStratNameData(strat_name_id: number) {
+  const stratNames = useStratNames([strat_name_id]);
+  return stratNames?.[0];
+}
+
+function StratNameField(
+  props: {
+    strat_name_id: number;
+    className?: string;
+  } & ItemInteractionProps,
+) {
+  /** Handling for stratigraphic name field */
+  const { strat_name_id, className, ...rest } = props;
+  const data = useStratNameData(strat_name_id);
+
+  const baseInteractionProps = useInteractionProps({ strat_name_id });
+
+  const coreProps = {
+    ...baseInteractionProps,
+    ...rest,
+  };
+
   let inner: any = h(Identifier, { id: strat_name_id });
-  const name = stratNameData?.strat_name_long;
+  const name = data?.strat_name_long;
   if (name != null) {
     inner = h("span.strat-name", name);
   }
 
-  const clickable = onClickItem != null;
-
-  const className = classNames({
-    clickable,
-  });
+  const clickable = isClickable(coreProps);
 
   return h(
     DataField,
@@ -363,8 +401,8 @@ function StratNameField({
     h(
       clickable ? "a" : "span",
       {
-        className,
-        onClick: (e) => onClickItem(e, { strat_name_id }),
+        className: classNames({ clickable }, className),
+        ...coreProps,
       },
       inner,
     ),
@@ -566,36 +604,32 @@ function enhanceLithologies(
 }
 
 export function ClickableText({
-  onClick,
   className,
-  children,
+  ...rest
 }: {
-  onClick: () => void;
   className?: string;
   children: ReactNode;
-}) {
+} & ItemInteractionProps) {
   /** An optionally clickable text element */
-  const tag = onClick != null ? "a" : "span";
-  return h(tag, { onClick, className }, children);
+  const clickable = isClickable(rest);
+  const tag = clickable ? "a" : "span";
+  return h(tag, { className: classNames(className, { clickable }), ...rest });
 }
 
 export function Identifier({
   id,
-  onClick,
   className,
+  ...rest
 }: {
   id: number | string;
-  onClick?: (id: number | string) => void;
   className?: string;
-}) {
+} & ItemInteractionProps) {
   /** An item that displays a numeric identifier, optionally clickable */
-  const _onClick = onClick != null ? () => onClick(id) : null;
-
   return h(
     ClickableText,
     {
-      onClick: _onClick,
       className: classNames("identifier", className),
+      ...rest,
     },
     id,
   );
@@ -607,8 +641,10 @@ type UnitInfo = {
   name?: string;
 };
 
-function UnitIDList({ units, onSelectUnit, showNames = false }) {
+function UnitIDList({ units, showNames = false }) {
   const unitsMap = useColumnUnitsMap();
+
+  const interactionManager = useInteractionManager();
 
   const extUnits: UnitInfo[] = useMemo(() => {
     const u1 = units.filter((d) => d != 0);
@@ -637,9 +673,17 @@ function UnitIDList({ units, onSelectUnit, showNames = false }) {
   return h(
     ItemList,
     { className: "units-list" },
-    extUnits.map((info) =>
-      h("span.item", h(UnitIdentifier, { ...info, onSelectUnit })),
-    ),
+    extUnits.map((info) => {
+      return h(
+        "span.item",
+        h(UnitIdentifier, {
+          ...info,
+          ...interactionManager?.interactionPropsForItem({
+            unit_id: info.unitID,
+          }),
+        }),
+      );
+    }),
   );
 }
 
@@ -647,21 +691,14 @@ function UnitIdentifier({
   unitID,
   colID,
   name,
-  onSelectUnit,
-}: UnitInfo & { onSelectUnit?: (unitID: number) => void }) {
-  const onClick = useMemo(() => {
-    if (onSelectUnit == null) return null;
-    return () => {
-      onSelectUnit(unitID);
-    };
-  }, [onSelectUnit]);
-
+  ...interactionProps
+}: UnitInfo & ItemInteractionProps) {
   if (name != null) {
     return h(
       ClickableText,
       {
         className: "unit-name",
-        onClick,
+        ...interactionProps,
       },
       name,
     );
@@ -669,9 +706,9 @@ function UnitIdentifier({
 
   return h(Identifier, {
     className: "unit-id",
-    onClick,
     key: unitID,
     id: unitID,
+    ...interactionProps,
   });
 }