+++ packages/data-components/src/index.ts
@@ -2,3 +2,5 @@ export * from "./components";
 export * from "./dz-spectrum";
 export * from "./field-locations";
 export * from "./location-info";
+export * from "./data-links";
+export * from "./utils";