+++ packages/data-provider/src/fetch.ts
@@ -5,11 +5,7 @@ import {
   StratName,
   UnitLong,
 } from "@macrostrat/api-types";
-import {
-  addQueryString,
-  joinURL,
-  useAPIResult,
-} from "@macrostrat/ui-components";
+import { addQueryString, joinURL } from "@macrostrat/ui-components";
 import crossFetch from "cross-fetch";
 import { feature } from "topojson-client";
 import { geoArea } from "d3-geo";
@@ -24,12 +20,16 @@ const defaultFetch = createScopedFetch("https://macrostrat.org/api/v2");
 
 export type ColumnStatusCode = "in process" | "active" | "obsolete";
 
-export interface ColumnFetchOptions {
+interface FetchBaseOptions {
+  // The fetch implementation to use
+  fetch?: (url: string, options?: RequestInit) => Promise<Response>;
+}
+
+export interface ColumnFetchOptions extends FetchBaseOptions {
   apiBaseURL?: string;
   projectID?: number;
   statusCode?: ColumnStatusCode | ColumnStatusCode[];
   format?: "geojson" | "topojson" | "geojson_bare";
-  fetch?: any;
 }
 
 export async function fetchAllColumns(
@@ -53,7 +53,9 @@ export async function fetchAllColumns(
   } else if (statusCode != null) {
     _statusCode = statusCode;
   }
-  args.statusCode = _statusCode;
+  if (_statusCode != null) {
+    args.status_code = _statusCode;
+  }
 
   if (projectID == null) {
     args = { ...args, all: true };
@@ -81,26 +83,6 @@ export async function fetchAllColumns(
   // Get JSON
   const data = await res.json();
   return postProcessColumns(columnProcessors[format](data));
-}
-
-export function useColumnFeatures({
-  apiRoute = "/columns",
-  status_code,
-  project_id,
-  format = "geojson",
-}) {
-  let all: boolean = undefined;
-  if (status_code == null && project_id == null) {
-    all = true;
-  }
-
-  const processor = columnProcessors[format];
-
-  return useAPIResult(
-    apiRoute,
-    { format, all, status_code, project_id },
-    processor,
-  );
 }
 
 function processGeoJSON(res) {
@@ -171,15 +153,17 @@ async function unwrapResponse(res) {
   return resData["success"]["data"];
 }
 
-export async function fetchLithologies(fetch = defaultFetch) {
+export async function fetchLithologies(opts: FetchBaseOptions = {}) {
+  const { fetch = defaultFetch } = opts;
   const res = await fetch("/defs/lithologies?all");
   return await unwrapResponse(res);
 }
 
 export async function fetchIntervals(
   timescaleID: number | null,
-  fetch = defaultFetch,
+  opts: FetchBaseOptions = {},
 ) {
+  const { fetch = defaultFetch } = opts;
   let url = `/defs/intervals`;
   if (timescaleID != null) {
     url += `?timescale_id=${timescaleID}`;
@@ -190,15 +174,17 @@ export async function fetchIntervals(
   return await unwrapResponse(res);
 }
 
-export async function fetchEnvironments(fetch = defaultFetch) {
+export async function fetchEnvironments(opts: FetchBaseOptions = {}) {
+  const { fetch = defaultFetch } = opts;
   const res = await fetch("/defs/environments?all");
   return await unwrapResponse(res);
 }
 
 export async function fetchRefs(
   refs: number[],
-  fetch = defaultFetch,
+  opts: FetchBaseOptions = {},
 ): Promise<MacrostratRef[]> {
+  const { fetch = defaultFetch } = opts;
   let url = `/defs/refs`;
   if (refs.length == 0) {
     return [];
@@ -210,8 +196,9 @@ export async function fetchRefs(
 
 export async function fetchStratNames(
   names: number[],
-  fetch = defaultFetch,
+  opts: FetchBaseOptions = {},
 ): Promise<StratName[]> {
+  const { fetch = defaultFetch } = opts;
   let url = `/defs/strat_names`;
   if (names.length == 0) {
     return [];
@@ -228,8 +215,9 @@ export type ColumnData = {
 
 export async function fetchUnits(
   columns: number[],
-  fetch = defaultFetch,
+  opts: FetchBaseOptions = {},
 ): Promise<ColumnData[]> {
+  const { fetch = defaultFetch } = opts;
   const params = new URLSearchParams();
   params.append("response", "long");
 
@@ -263,8 +251,9 @@ export async function fetchUnits(
 
 export async function fetchColumnUnits(
   col_id: number,
-  fetch = defaultFetch,
+  opts: FetchBaseOptions = {},
 ): Promise<ColumnData> {
+  const { fetch = defaultFetch } = opts;
   const params = new URLSearchParams();
   params.append("response", "long");
   params.append("col_id", col_id.toString());