+++ packages/column-views/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2026-01-31
+
+- Moved `MacrostratDataProvider` and data fetchers to
+  `@macrostrat/data-provider` for better modularity.
+- Standardize approach to clickable/linkable data items in `UnitDetailsPanel`,
+  using a new `MacrostratInteractionProvider` from
+  `@macrostrat/data-components`.
+
 ## [3.0.3] - 2026-01-29
 
 - Change layout of `package.json`